+++ docs/history.rst
@@ -4,7 +4,27 @@
 History
 =======
 
+Version 0.4.0 - Unreleased
+--------------------------
+
+* Renamed project from "html-classes-linter" to "chalumo";
+* Finished reformater class;
+
+
+Version 0.3.0 - Unreleased
+--------------------------
+
+Finished basic structure implementation with CLI, parser, pre processor, differ and
+tests.
+
+
+Version 0.2.0 - Unreleased
+--------------------------
+
+Just startint to implement parsing.
+
+
 Version 0.1.0 - Unreleased
 --------------------------
 
-* First commit.
+* First commit, it was started with name "html-classes-linter";